refactored to use concerns
